# .well-known/openapi.yaml
openapi: 3.0.1
info:
    title: 答案之书
    description: 答案之书是一本神奇的书，能够根据你的问题或者困惑生成对应的解答，为你指点迷津。
    version: "v1"
servers:
    - url: http://127.0.0.1:8082
paths:
    /Introduce_the_Book:
        post:
            operationId: IntroduceBook
            summary: 给出答案之书的介绍和使用说明。
            responses:
                "200":
                    description: 回执
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/OutputMessage"
            x-return-raw:
                description: 标记当前API请求的结果是否经过大模型，当name的值不空时，从插件结果中获取具体字段，类型必须string，直接返回，而不经过大模型。
                mediaType: application/json
                name: message
                type: string
    /Get_Answer_of_Number:
        post:
            operationId: GetAnswerofNumber
            summary: 打开答案之书特定一面，查看对应内容并且给出解释
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/InputNums"
            responses:
                "200":
                    description: 答案数据返回
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/OutputMessage"
            x-return-raw:
                description: 标记当前API请求的结果是否经过大模型，当name的值不空时，从插件结果中获取具体字段，类型必须string，直接返回，而不经过大模型。
                mediaType: application/json
                name: message
                type: string
    /Shuffle_Book:
        post:
            operationId: ShuffleBook
            summary: 随机刷新或重置答案之书
            responses:
                "200":
                    description: 回执
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/OutputMessage"
            x-return-raw:
                description: 标记当前API请求的结果是否经过大模型，当name的值不空时，从插件结果中获取具体字段，类型必须string，直接返回，而不经过大模型。
                mediaType: application/json
                name: message
                type: string
    /Get_Question_and_Find_Three_Answers:
        post:
            operationId: GetQuestionandFindThreeAnswers
            summary: 在用户有疑惑或者问题的时候，随机翻开答案之书，抽取三个回复
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/InputQuestion"
            responses:
                "200":
                    description: 回执
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/OutputMessage"
            x-return-raw:
                description: 标记当前API请求的结果是否经过大模型，当name的值不空时，从插件结果中获取具体字段，类型必须string，直接返回，而不经过大模型。
                mediaType: application/json
                name: message
                type: string

components:
    schemas:
        InputNums:
            type: object
            required: [number]
            properties:
                number:
                    type: integer
                    description: 获取输入的一个数字
        InputQuestion:
            type: object
            required: [question]
            properties:
                question:
                    type: string
                    description: 想要咨询或者感到疑惑的问题
        InputQuestionandNums:
            type: object
            required: [number, question]
            properties:
                number:
                    type: integer
                    description: 获取输入的一个数字
                question:
                    type: string
                    description: 想要咨询或者感到疑惑的问题
        OutputMessage:
            type: object
            required: [message]
            properties:
                message:
                    type: string
                    description: 回复信息